replace tweet() example; closes #73, closes #84
Instead of writing a tweet() function where we need to explain
that nothing is linked with Twitter and possibly need to explain
Python, we rather collaboratively create a taco recipe cookbook
where we only write markdown files.

This is inspired by https://github.com/sinker/tacofancy.

One addition to the exercise is that we first create an issue
before we do any coding to collect feedback and also to
get used to auto-closing issues.

The automated testing checks whether all markdown files
contain the string "taco".