Add basic motivation
The episode goes straight to technical stuff, even in the index.  I
think it needs some basic motivation of where we came from and what we
get out of it.  I think these four levels provide a good starting
point.

I'm putting them here instead of new episode to minimize the changes,
but it can be reorganized as needed.